move distributor context to swarm
