+++ jobmon_client/src/jobmon/client/swarm/workflow_run.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 
 import ast
 from collections import deque
+from contextlib import nullcontext
 from datetime import datetime
 import logging
 import time
@@ -19,6 +20,7 @@ from typing import (
 )
 
 from jobmon.client.array import Array
+from jobmon.client.distributor_context import DistributorContext
 from jobmon.client.swarm.swarm_array import SwarmArray
 from jobmon.client.swarm.swarm_task import SwarmTask
 from jobmon.client.task_resources import TaskResources
@@ -84,7 +86,6 @@ class WorkflowRun:
         fail_after_n_executions: int = 1_000_000_000,
         status: Optional[str] = None,
         requester: Optional[Requester] = None,
-        remote_distributor: bool = False,
     ) -> None:
         """Initialization of the swarm WorkflowRun."""
         self.workflow_run_id = workflow_run_id
@@ -131,10 +132,6 @@ class WorkflowRun:
         if requester is None:
             requester = Requester.from_defaults()
         self.requester = requester
-
-        # Determines whether a local distributor is actively managing this swarm or if it's
-        # managed by a remote distributor instance
-        self.remote_distributor = remote_distributor
 
         # This signal is set if the workflow run receives a resume
         self._terminated = False
@@ -433,9 +430,11 @@ class WorkflowRun:
 
     def run(
         self,
-        distributor_alive_callable: Callable[..., bool],
         seconds_until_timeout: int = 36000,
         initialize: bool = True,
+        remote_distributor: bool = False,
+        cluster_name: str = '',
+        distributor_startup_timeout: int = 180,
     ) -> None:
         """Take a concrete DAG and queue al the Tasks that are not DONE.
 
@@ -459,121 +458,130 @@ class WorkflowRun:
             seconds_until_timeout: how long to block while waiting for the next task to finish
                 before raising an error.
             initialize: whether to initialize (update WorkflowRun to RUNNING and set fringe)
-
+            remote_distributor: whether to connect to a remote distributor or create a local
+                one
+            cluster_name: the cluster to run against, if creating a local distributor
+            distributor_startup_timeout: how long to wait for a local distributor to register
+                before raising an error
         Return:
             None
         """
-        try:
-            if initialize:
-                logger.info(f"Executing Workflow Run {self.workflow_run_id}")
-                self.set_initial_fringe()
-                self._update_status(WorkflowRunStatus.RUNNING)
-
-            time_since_last_full_sync = 0.0
-            total_elapsed_time = 0.0
-            terminating_states = [
-                WorkflowRunStatus.COLD_RESUME,
-                WorkflowRunStatus.HOT_RESUME,
-            ]
-
-            while self.active_tasks:
-                # Expire the swarm after the requested number of seconds
-                if total_elapsed_time > seconds_until_timeout:
-                    raise RuntimeError(
-                        f"Not all tasks completed within the given workflow timeout length "
-                        f"({seconds_until_timeout} seconds). Submitted tasks will still run, "
-                        f"but the workflow will need to be restarted."
-                    )
-
-                # check that the distributor is still alive
-                if not distributor_alive_callable():
-                    raise DistributorNotAlive(
-                        "Distributor process unexpectedly stopped. Workflow will error."
-                    )
-
-                # If the workflow run status was updated asynchronously, terminate
-                # all active task instances and error out.
-                if self.status in terminating_states:
-                    logger.warning(
-                        f"Workflow Run set to {self.status}. Attempting graceful shutdown."
-                    )
-                    # Active task instances will be set to "K", the processing loop then
-                    # keeps running until all of the states are appropriately set.
-                    self._terminate_task_instances()
-
-                # if fail fast and any error
-                if self.fail_fast and self._task_status_map[TaskStatus.ERROR_FATAL]:
-                    logger.info("Failing after first failure, as requested")
-                    break
-
-                # fail during test path
-                if self._n_executions >= self._val_fail_after_n_executions:
-                    raise WorkflowTestError(
-                        f"WorkflowRun asked to fail after {self._n_executions} "
-                        "executions. Failing now"
-                    )
-
-                # process any commands that we can in the time allotted
-                loop_start = time.time()
-                time_till_next_heartbeat = self._workflow_run_heartbeat_interval - (
-                    loop_start - self._last_heartbeat_time
-                )
-                if self.status == WorkflowRunStatus.RUNNING:
-                    self.process_commands(timeout=time_till_next_heartbeat)
-
-                # take a break if needed
-                loop_elapsed = time.time() - loop_start
-                if loop_elapsed < time_till_next_heartbeat:
-                    sleep_time = time_till_next_heartbeat - loop_elapsed
-                    time.sleep(sleep_time)
-                    loop_elapsed += sleep_time
-
-                # then synchronize state
-                if time_since_last_full_sync > self.wedged_workflow_sync_interval:
-                    time_since_last_full_sync = 0.0
-                    self.synchronize_state(full_sync=True)
-                else:
-                    time_since_last_full_sync += loop_elapsed
-                    self.synchronize_state()
-
-                total_elapsed_time += time.time() - loop_start
-
-        # user interrupt
-        except KeyboardInterrupt:
-            logger.warning("Keyboard interrupt raised")
-            confirm = input("Are you sure you want to exit (y/n): ")
-            confirm = confirm.lower().strip()
-
-            if confirm == "y":
-                self._update_status(WorkflowRunStatus.STOPPED)
-                raise
-            else:
-                logger.info("Continuing jobmon...")
-                seconds_until_timeout = int(seconds_until_timeout - loop_elapsed)
-                self.run(
-                    distributor_alive_callable, seconds_until_timeout, initialize=False
-                )
-
-        # unexpected errors. raise
-        except Exception as e:
-            try:
-                self._update_status(WorkflowRunStatus.ERROR)
-            except TransitionError as trans:
-                logger.warning(trans)
-            raise e
-
-        # no more active tasks
+        if not remote_distributor:
+            context = DistributorContext(cluster_name=cluster_name,
+                                         workflow_run_id=self.workflow_run_id,
+                                         timeout=distributor_startup_timeout,
+                                         logger=logger)
         else:
-            # check if done
-            if len(self.tasks) == len(self._task_status_map[TaskStatus.DONE]):
-                logger.info("All tasks are done")
-                self._update_status(WorkflowRunStatus.DONE)
+            context = nullcontext()
+            context.alive = lambda: True
+        if initialize:
+            logger.info(f"Executing Workflow Run {self.workflow_run_id}")
+            self.set_initial_fringe()
+            self._update_status(WorkflowRunStatus.RUNNING)
 
+        time_since_last_full_sync = 0.0
+        total_elapsed_time = 0.0
+        terminating_states = [
+            WorkflowRunStatus.COLD_RESUME,
+            WorkflowRunStatus.HOT_RESUME,
+        ]
+        with context:
+            while self.active_tasks:
+                try:
+                    # Expire the swarm after the requested number of seconds
+                    if total_elapsed_time > seconds_until_timeout:
+                        raise RuntimeError(
+                            f"Not all tasks completed within the given "
+                            f"workflow timeout length "
+                            f"({seconds_until_timeout} seconds). "
+                            f"Submitted tasks will still run, "
+                            f"but the workflow will need to be restarted."
+                        )
+
+                    # check that the distributor is still alive
+                    if not context.alive():
+                        raise DistributorNotAlive(
+                            "Distributor process unexpectedly stopped. Workflow will error."
+                        )
+
+                    # If the workflow run status was updated asynchronously, terminate
+                    # all active task instances and error out.
+                    if self.status in terminating_states:
+                        logger.warning(
+                            f"Workflow Run set to {self.status}. Attempting graceful shutdown."
+                        )
+                        # Active task instances will be set to "K", the processing loop then
+                        # keeps running until all of the states are appropriately set.
+                        self._terminate_task_instances()
+
+                    # if fail fast and any error
+                    if self.fail_fast and self._task_status_map[TaskStatus.ERROR_FATAL]:
+                        logger.info("Failing after first failure, as requested")
+                        break
+
+                    # fail during test path
+                    if self._n_executions >= self._val_fail_after_n_executions:
+                        raise WorkflowTestError(
+                            f"WorkflowRun asked to fail after {self._n_executions} "
+                            "executions. Failing now"
+                        )
+
+                    # process any commands that we can in the time allotted
+                    loop_start = time.time()
+                    time_till_next_heartbeat = self._workflow_run_heartbeat_interval - (
+                        loop_start - self._last_heartbeat_time
+                    )
+                    if self.status == WorkflowRunStatus.RUNNING:
+                        self.process_commands(timeout=time_till_next_heartbeat)
+
+                    # take a break if needed
+                    loop_elapsed = time.time() - loop_start
+                    if loop_elapsed < time_till_next_heartbeat:
+                        sleep_time = time_till_next_heartbeat - loop_elapsed
+                        time.sleep(sleep_time)
+                        loop_elapsed += sleep_time
+
+                    # then synchronize state
+                    if time_since_last_full_sync > self.wedged_workflow_sync_interval:
+                        time_since_last_full_sync = 0.0
+                        self.synchronize_state(full_sync=True)
+                    else:
+                        time_since_last_full_sync += loop_elapsed
+                        self.synchronize_state()
+
+                    total_elapsed_time += time.time() - loop_start
+                # user interrupt
+                except KeyboardInterrupt:
+                    logger.warning("Keyboard interrupt raised")
+                    confirm = input("Are you sure you want to exit (y/n): ")
+                    confirm = confirm.lower().strip()
+
+                    if confirm == "y":
+                        self._update_status(WorkflowRunStatus.STOPPED)
+                        raise
+                    else:
+                        logger.info("Continuing jobmon...")
+
+                # unexpected errors. raise
+                except Exception as e:
+                    try:
+                        self._update_status(WorkflowRunStatus.ERROR)
+                    except TransitionError as trans:
+                        logger.warning(trans)
+                    raise e
+
+            # no more active tasks, loop terminated without an exception raised
             else:
-                if self.status in terminating_states:
-                    self._update_status(WorkflowRunStatus.TERMINATED)
+                # check if done
+                if len(self.tasks) == len(self._task_status_map[TaskStatus.DONE]):
+                    logger.info("All tasks are done")
+                    self._update_status(WorkflowRunStatus.DONE)
+
                 else:
-                    self._update_status(WorkflowRunStatus.ERROR)
+                    if self.status in terminating_states:
+                        self._update_status(WorkflowRunStatus.TERMINATED)
+                    else:
+                        self._update_status(WorkflowRunStatus.ERROR)
 
     def set_initial_fringe(self) -> None:
         """Set initial fringe."""
