+++ jobmon_distributor/src/jobmon/distributor/distributor_instance.py
@@ -115,6 +115,9 @@ class DistributorInstance:
         else:
             self.requester = requester
 
+        # TODO: prior distributor transitions WFR to instantiated, prevent reaping
+        #  check if still necessary
+
     @property
     def _next_report_increment(self) -> float:
         return self._heartbeat_report_by_buffer * self._task_instance_heartbeat_interval
